Generate more stack maps during native debugging.
Generate extra stack map at the start of each java statement.
The stack maps are later translated to DWARF which allows
LLDB to set breakpoints and view local variables.

The extra stack maps are generated by new HNativeDebugInfo
pseudo-instruction.  The presence of this instruction might
interfere with some tests which compare the list of instructions
to known list.  So do not run tests with --generate-debug-info
unless specifically requested by the test (it is required by CFI).

Change-Id: If00ab875513308e4a1399d1e12e0fe8934a6f0c3