[WIP] Add class profiling
Class profiling is a way to keep track of which classes are loaded
and verified. From here the compiler can use this to generate a
smaller app image. The runtime can also use this to avoid
verification next run.

Collecting the class profile is fast, ~2ms on maps. We can probaly
run this often with exponential backoff for all apps running if we
want verification info for more than just the first few seconds
worth of loaded classes.

Bug: 22858531
Change-Id: I9e62240a052703e2c4ac768996bfe33140743ffa