ART: Fixup JDWP tests script
The tests have problems on the host because of boot images. Vogar
by default runs in 32-bit mode, but the art script, used for forked
debuggee processes, by default runs in 64-bit. The JDWP forks also
do not set a boot classpath. With that mismatch, when the tests
are run in 32-bit mode on a 64-bit enabled host, the forked processes
will die.

Check whether a --variant parameter was passed to the test (and if
not assume it's x32). If the parameter is (x)32, then adapt the
debuggee command to ensure the forked runtime also runs in 32-bit
mode and finds the image from the controller.

Change-Id: Ic22477e33702e1a7c1ca94466bf02df4cb201723