Remove an extra implicit null check in the ARM codegen.
The code generated for integer ArraySet on ARM used to
contain two implicit null checks for the array.  This has
not been an actual issue so far, as ArraySet instructions
have never been using implicit null checks.

Note: This CL comes without a regression test, as the code
path in question is not used (yet).

Change-Id: I7dccb02cf7ac2f7d8fd1676b03e0b394701fbe3f