ART: Optimize out redundant NewInstances of String
NewInstance of String creates an empty String object before it is
replaced by the result of a StringFactory call (String.<init>). If
the empty object is never used prior to the call, it can be safely
removed (replaced with null in this case).

We do not remove the instruction if:
 - it has a real use, e.g. comparison or (catch) phi, or
 - we are compiling with --debuggable and there is an environment use.
If removed and execution deoptimizes before the StringFactory call,
the interpreter will see String.<init> being called on a null object.
Since the verifier guarantees that the call was made on new-instance
in the input bytecode (b/26579108), the interpreter can safely assume
that it was optimized out rather than throw NullPointerException.

Results from compiling boot.oat (without --debuggable):
 - total # of instances:   478
 - removed:                466   97%
 - has real use:            12    3%

Results from APKs:
 - Google Maps: 425/425 = 100% removed
 - Google Docs: 812/812 = 100% removed

Change-Id: I9fdfc50e9dea6299a7327f94327cdfd2b2538079