ART: Optimize out redundant NewInstances of String
NewInstance of String creates an empty String object before it is
replaced by the result of a StringFactory call (String.<init>). If
the empty object is never used prior to the call, it can be safely
removed.

We do not remove the instruction if:
 - it has a real use, e.g. comparison or (catch) phi, or
 - it is in the environment of Deoptimize, since the interpreter
   assumes NewInstance is a non-null Object, or
 - we are compiling with --debuggable and there is an environment use

Results from compiling boot.oat (without --debuggable):
 - total # of instances:   478
 - removed:                466   97%
 - has real use:            12    3%
 - deoptimized:              0    0%

Results from APKs:
 - Google Maps: 425/425 = 100% removed
 - Google Docs: 812/812 = 100% removed

Change-Id: I9fdfc50e9dea6299a7327f94327cdfd2b2538079