Emit some additional debug info for java.lang.String
The value part of java.lang.String is not a proper ArtField so the
default debug info writing code don't emit debug info about that. This
CL add logic to emit debug info for a new member what represents the
first element of the underlying string data what can be used by the
consumer of the debug info to display the value of the string.

Change-Id: Ibc11b143e79aa8866000d02922bcd6069cfcd4c2