Put the perf map file to the application directory
The perf map file is used in the jit compiler to emit the symbol name
for the start address of each jitted function. Put this file into the
application directory if it have been specified so it can be created
by the application user. We still put it into /data if the application
directory isn't specified.

Change-Id: Icfecd4bdab94ffc528ec218f3ac2b872fbdacf37