ART: Adjust liveness of non-materialized condition inputs
The optimization of not materializing conditions does not adjust the
use positions of the Condition's inputs and instead relies on the
fact that the register allocator will never insert a ParallelMove
between the Condition and its user. This will no longer hold with
the introduction of HSelect and therefore needs to be accounted for.

Intuitively, the inputs of a non-materialized condition are in fact
used by the user of the condition, not the condition itself, and
therefore need to be kept live until then.

Change-Id: Id703c154b20ab861241ae5c715a150385d3ff621