Enable --generate-mini-debug-info by default.
This generates the bare minimum of data necessary to support backtraces
in libunwind (that is, LZMA-compressed symtab and debug_frame sections).
Other non-essential debugging info is excluded (e.g. no line numbers).

Symtab contains java method names (with class names, but without
parameter names and types to save a little extra space).

The compression is done on background thread during final I/O-bound
stages of dex2oat when the CPU would be otherwise underutilized.
This helps us hide some (but not all) of the time overhead.

Results on boot.oat compilation time on the device:
 * 3.5% increase in compilation time (from 30.8s to 31.9s)
 * 2.5% increase of boot.oat size (from 75MB to 77MB)

Change-Id: I0b005ab53d7696032d0a35c13eb25a5ce85cfb7d