interpreter: Fix proxy method invocation with access checks enabled.
Interpret with access checks uses type information stored by a dex file,
so we need to get the original interface method (which is guaranteed
to be dex file backed) in that case.

Does not change method invoke target, just the method used for checks.

(The existing 044-proxy test already covers this, so no regression test
necessary)

Bug: 26846861
Bug: 22414682
Change-Id: I28c3707e38c1f69ea9d3660f68136f688122ac4e

Remove tests from blacklist

Change-Id: I6813e0aa4f1ebf31b4a646a87b28488a79495c29