Allow oat file updates during dex loading.
If the caller passes an oat_location then attempt to update the loaded
dex file (i.e. compile or relocate). That's basically the path taken
when loading classes with DexClassLoader.

It the oat_location is null, we will not block to generate or update any
oat files, even if they are out of date. This is the path taken by
PathClassLoader.

Bug: 26719109
Change-Id: I9650207ed115e60239ad809fb45f206df85f37b3