Improve handling of oat file updates during dex loading
Prior to this CL OatFileManger will always attempt to make the oat file
up to date (compile or relocate). The default oat location when not
specied was dalvik-cache. However, in most cases the caller would not
have rights to write there resulting in confusing errors being logged.
The only cases when we have rights are 1) tests and 2) system server
(which shouldn't need to do this since everything is already compiled).

The behaviour changes as fallows:

If the caller passes an oat_location then attempt to update the loaded
dex file (i.e. compile or relocate). That's basically the path taken
when loading classes with DexClassLoader.

It the oat_location is null, we will not block to generate or update any
oat files, even if they are out of date. This is the path taken by
PathClassLoader.

--no-prebuild tests pass -Xforce-oat-update-at-load-time to the runtime
which will force the oat file update. This preserves the current
semantic of the tests.

Bug: 26719109
Change-Id: I9650207ed115e60239ad809fb45f206df85f37b3