[RFC] Enhance Induction Variable Analysis
This is something I have been playing with. It adds support for
limited type conversion (int->byte and int->short) to handle loops
that have byte or short induction variables.  Ensure that the bounds
are in range.

The goal is to allow BCE to handle:
    a[0] = 0;
    for (byte i = 1; i < 120; i++) {
      a[i] = 0;
    }

The loop needs to be peeled to allow GVN to pull the NullCheck from the
loop and let BCE move the BoundsCheck. We can talk about automatic
peeling if there is an interest.

I don't like how I did the bounds checking code.  Is there some better
way to handle it? I wanted the range checking to be more specific to IV
with a HTypeCheck, but couldn't find it.

There are some benchmarks (such Quadrant) that have for loops using
FP values that are integral.  I would like to use IVA to be able
to compute trip counts, but IVA only supports integers for lower,
upper, and strides.  Enhance it to support FP values as well, IFF
the FP values all have integer values.  This ensures that we don't
have to worry about rounding errors.

I would also like to change IVA to make the types and LookupInfo public,
so that they can be used by all optimizations, not just IVR. Comments on
that?

Change-Id: I6d8843f2934677735c05939caa621be8f60fba4a
Signed-off-by: Mark Mendell <mark.p.mendell@intel.com>