ART: Add classlinker_classes_lock_ when we modify class members, for race condition.
GC thread would visit mismatched data of class members
when another thread is doing DefineClass, and the VM will crash.

The class members should not be modified while doing
VisitClassRoots of GC. Add the lock 'classlinker_classes_lock_'
to protect the class data when we want to modify the class.

Change-Id: I5f47d1d61b005e60704873fc87edf703d31f73d7