GSS/SS fix potential issue of fallback space.
For GSS/SS, fallback space is used when fail to copy object to to_space
or promote space. in this situation, bitmap of fallback space should be
marked.
This patch fix an issue that mark bitmap was not marked for fallback
space. and also if collect_from_space_only, both should not marked.

In current GSS/SS implementation, it is hard to copy object into
fallback space. but this potential issue needs to be fixed to support
parallel copying.

Change-Id: Ic17f58eda7892521bbafc54684e075446ad4d535
Signed-off-by: Lin Zang <lin.zang@intel.com>