Refine the return code of GetDexOptNeeded
Make explicit why DexOpt is not needed by adding more return codes to
GetDexOptNeeded.

extract-only or profile-guide oat files are considered up to date from
runtime perspective as they don't necessary need (re)compilation or
relocation. However, it is useful to return a more refined code to the
caller so that they can decide whether or not that's good enough.

For example, the package manager might decide to still compile a
previous extract-only and during profile guide compilation we should
always recompile even if we have an oat file.

This change also introduces a new key in the oat header
kProfileGuideCompiledKey. The distinction is needed in order to avoid
recompilation of a previous fully compiled file during profile guide
compilation analysis.

Bug: 27189430
Change-Id: Icd9794b1df6f6e21242e1dd1d5b5d064963dbbb7