Refine the return code of GetDexOptNeeded
Add OatIsExtractOnly return code to GetDexOptNeeded.

Extract only oat files are considered up to date from runtime
perspective as they don't necessary need (re)compilation or relocation.
However, it is useful to return a more refined code to the caller so
that they can decide whether or not just an extracted oat file is good
enough.

Bug: 27189430
Change-Id: Icd9794b1df6f6e21242e1dd1d5b5d064963dbbb7