[WIP] ART: Add image compilation determinism test
Refactor image_test to allow compiling an image and do some work
with it. Use this to, in a loop, build and checksum the boot image,
and see whether any changes are detected.

Requires some changes to the test and infrastructure, so we can
compile with the same initial conditions, and prepared for a
deterministic run.

Flaky failures, indicates we're still missing something somewhere.
Working on a principled way of getting the data out.

Bug: 26687569
Change-Id: I6063dca9a1909d226cd79d4981bd22ec56adfe76