[WIP] Optimizing: Improve const-string code generation.
For strings in the boot image, use either direct pointers
or pc-relative addresses. For other strings, use PC-relative
access to the dex cache arrays for AOT and direct address of
the string's dex cache slot for JIT.

DONE: x86, x86-64, arm64.
TODO: arm;
TODO: debug arm64 - it looks like we're exposing a register
allocator bug in
  long com.android.server.usage.UsageStatsXml.parseBeginTime(java.io.File)
where we use the current ArtMethod* instead of the result of the
File.getName() call.

Bug: 26884697
Change-Id: Iec7266ce67e6fedc107be78fab2e742a8dab2696