WIP ART: Improve JitProfile performance in mterp
ART currently requires two profiling-related things from the interpreters:
hotness updates and OSR switch checks.  The hotness updates previously used
the existing instrumentation framework - which is flexible, but quite
heavyweight.  For most things, the instrumentation framework overhead is
acceptable, but because we do a hotness update on every backwards branch
the overhead is unacceptable.  Prior to this CL, branch profiling dominates
interpreter cost.

The most efficient hotness update mechanism would simply bypass the
instrumentation framework.  However, from a code health point of view, this
would be somewhat unclean.  To avoid introducing a parallel instrumentation
mechanism, this CL adds a new instrumentatation "event" using the existing
framework - a batch update of hotness count.  This will enable us to locally
update hotness, but only periodically call out to the instrumentation
framework to amortize the cost.

The existing "Branch" event will be retained, and in the future is expected
to be used to collect information to drive better optimizations.  The
expectation is that a method will start off with hotness profiling, and
once it reaches a hotness threshold that suggests it is worth optimizing,
we turn on the expsive full branch profiling long enough to collect useful
information.

Once the method is sent to the optimizer for compilation, hotness profiling
would resume in order to detect candidates for on-stack replacement.

The CL also includes some minor ARM assemby cleanup changes.

Change-Id: I50b5b8c5ed8ebda3c8b4e65d27ba7393c3feae04