WIP: Generate native debug info for AOTed methods in boot.oat.
Force-JITing executed methods in boot.oat as native-debuggable
has non-negligible performance impact. Instead of doing that,
generate DWARF data based on stackmaps in the AOTed boot.oat.

This means it is not possible to set breakpoints or step
through the framework code under LLDB, however we still get
full backtrace with accurate java line numbers.

Note that the boot.oat file does not get affected. The DWARF
generation happens when we execute app as native-debuggable.

Change-Id: I73ff6f43a3a0e5a0907af353f74f04e9b0129590