Fix valgrind errors with MemMap::Sync().
This fixes one error in valgrind-test-art-host-gtest-oat_test and one
error in valgrind-test-art-host-gtest-image_test32.

Valgrind doesn't like it if an address range contains noaccess or
uninitialized memory is passed to msync(). Temporarily lift the oaccess
protection of the lower-end redzone because msync accepts page-aligned
base addresss only and exclude the higher-end noaccess redzone from the
range.

Bug: 27552451
Bug: 27384445
Change-Id: I8ccbd04c62eb30f6c6d5c732f1eb254fa09a417a