Patch method entry points only if boot.oat is non-debuggable.
Related to https://android-review.googlesource.com/#/c/206134/
and https://android-review.googlesource.com/#/c/201383/

boot.oat has been reverted back to debuggable, therefore we do not
need to do the method entry point patching. Instead of reverting
the code, make it conditional on Runtime::IsDebuggable().

This is needed as the patching interferes with performance of
native debugging. Added TODO to dex2oat to keep note of this.

Change-Id: I66c420a567e21073253f63dc2cc632f0b853d87d