Minor optimizations of debug::WriteCFISection.
The method might be passed method infos without CFI in some cases.
Use the sorting phase as an chance to filter them out.
This makes sure we do not allocate memory and sort methods if there
is in fact to no work to do.

Also change the sort to stable - for the sake of determinism.

Change-Id: I97d57d77e8b709d0d49d6971f66b955efcbb57b0