Complete the 1.8 unsafe memory fences tests.
Rationale:
The tests for the fences were still TODO, and this
change list completes them. Java does not quite allow
for a weak memory order on just the signal, but this
is as close as a real unit test as I can think of.
Also added a test suggested by Hans.

BUG=26264765

Change-Id: I2dcb9b02fa63f72daf4c39b3744ff2010888713e