Revert "DWARF: Compensate for compiler's off-by-one instruction."
This reverts commit 252fa90d25eb3fa9c4189304d021533c9657fea7.

The trick works great for non-debuggable oat files and line numbers
for debuggable oat files will also be accurate.
    
However, call-sites in debuggable files are not adjusted so we
might still get the local variable mapping for the wrong PC.
    
So we still need to teach LLDB about the off-by-one convention,
or we have to propagate more data through the compiler.

Change-Id: Ic60891027e1b17f3453d34058f273e31d9113b81