MIPS64: Test LoadConst32/LoadConst64.
Most of the LoadConst64() tests are for individual constants. A future
revision of the test is planned which will actually execute the
LoadConst64() generated code, and verify that the constant value loaded
is correct. When that happens the "large" tests for LoadConst32(),
LoadConst64() will be broken up into tests for individual constant
values.

Change-Id: Ie1ec050e65c666168a9ea582bae9fda97e4cf36a