Recognize and simplify zero-extension array reads.
Made Instruction Simplifier recognize ArrayGet & 0xFFFF as zero
extension memory reads. Similar improvement will be done with
FieldGet instructions.

GVN is fixed to distinguish sign- and zero- reads.
Load-Store-Elimination is fixed to zero-extend the source value
in place of the eliminated load if it was unsigned.

This patch makes use of the optimization only for x86 and x86_64.
The other code generators can be fixed in the same way.

Change-Id: I65630d864868b58ba7999c8a0acc81ba9b538c3c
Signed-off-by: Yevgeny Rouban <yevgeny.y.rouban@intel.com>