Use compiler filter to determine oat file status.
Record the compiler filter in the oat header. Use that to determine
when the oat file is up-to-date with respect to a target compiler
filter level.

New xxx-profile filter levels are added to specify if a profile should
be used instead of testing for the presence of a profile file.

This change should allow for different compiler-filters to be set for
different package manager use cases.

Bug: 27689078

(cherry picked from commit 5629d94d3f2e1d63ca71a67ca1599c830cc66aec)

Change-Id: Id6706d0ed91b45f307142692ea4316aa9713b023