Liveness: Handle non-first use equal to output
If the use of instruction is not its first input but
equal to first input and instruction requires the first
input is the same as output we can use the same position for
as we would be first input. It helps to avoid redundant movs on x86
for instructions like mul/add if both inputs are the same.

Change-Id: I83b78fdb86c887e4645ccd7f58a7e007436597b7
Signed-off-by: Serguei Katkov <serguei.i.katkov@intel.com>